+++ test/test_33_compare_predictions.rb
@@ -29,7 +29,7 @@ class Test33ComparePredictions < Test::Unit::TestCase
   def test_scenario1
     
     data = [
-           ['data/spam.csv', {"fields"=> {"000001"=> {"optype": "text", "term_analysis"=> {"case_sensitive"=> true, "stem_words"=> true, "use_stopwords"=> false, "language"=> "en"}}}}, {"Type"=> "ham", "Message"=> "Mobile call"}, 'Cluster 7', 0.36637],
+           ['data/spam.csv', {"fields"=> {"000001"=> {"optype": "text", "term_analysis"=> {"case_sensitive"=> true, "stem_words"=> true, "use_stopwords"=> false, "language"=> "en"}}}}, {"Type"=> "ham", "Message"=> "Mobile call"}, 'Cluster 0', 0.25],
            ['data/spam.csv', {"fields" => {"000001" => {"optype" => "text", "term_analysis" => {"case_sensitive" => true, "stem_words" => true, "use_stopwords" => false}}}}, {"Type" => "ham", "Message" => "A normal message"}, 'Cluster 0', 0.5],
            ['data/spam.csv', {"fields": {"000001"=> {"optype"=> "text", "term_analysis"=> {"case_sensitive"=> false, "stem_words"=> false, "use_stopwords"=> false, "language"=> "en"}}}}, {"Type"=> "ham", "Message"=> "Mobile calls"}, 'Cluster 0', 0.5],
            ['data/spam.csv', {"fields"=> {"000001"=> {"optype"=> "text", "term_analysis"=> {"case_sensitive"=> false, "stem_words"=> false, "use_stopwords"=> false, "language"=> "en"}}}}, {"Type"=> "ham", "Message"=> "A normal message"}, 'Cluster 0', 0.5],
@@ -37,7 +37,7 @@ class Test33ComparePredictions < Test::Unit::TestCase
            ['data/spam.csv', {"fields"=> {"000001"=> {"optype"=> "text", "term_analysis"=> {"case_sensitive"=> false, "stem_words"=> true, "use_stopwords"=> true, "language"=> "en"}}}}, {"Type"=> "ham", "Message"=> "A normal message"}, 'Cluster 1', 0.36637],
            ['data/spam.csv', {"fields"=> {"000001"=> {"optype"=> "text", "term_analysis"=> {"token_mode"=> "full_terms_only", "language"=> "en"}}}}, {"Type"=> "ham", "Message"=> "FREE for 1st week! No1 Nokia tone 4 ur mob every week just txt NOKIA to 87077 Get txting and tell ur mates. zed POBox 36504 W45WQ norm150p/tone 16+"}, 'Cluster 0', 0.5],
            ['data/spam.csv', {"fields"=> {"000001"=> {"optype"=> "text", "term_analysis"=> {"token_mode"=> "full_terms_only", "language"=> "en"}}}}, {"Type"=> "ham", "Message"=> "Ok"}, 'Cluster 0', 0.478833312167],
-           ['data/spam.csv', {"fields"=> {"000001"=> {"optype"=> "text", "term_analysis"=> {"case_sensitive"=> true, "stem_words"=> true, "use_stopwords"=> false, "language"=> "en"}}}}, {"Type"=> "", "Message"=> ""}, 'Cluster 1', 0.5],
+           ['data/spam.csv', {"fields"=> {"000001"=> {"optype"=> "text", "term_analysis"=> {"case_sensitive"=> true, "stem_words"=> true, "use_stopwords"=> false, "language"=> "en"}}}}, {"Type"=> "", "Message"=> ""}, 'Cluster 6', 0.5],
            ['data/diabetes.csv', {"fields"=> {}}, {"pregnancies"=> 0, "plasma glucose"=> 118, "blood pressure"=> 84, "triceps skin thickness"=> 47, "insulin"=> 230, "bmi"=> 45.8, "diabetes pedigree"=> 0.551, "age"=> 31, "diabetes"=> true}, 'Cluster 3', 0.5033378686559257],
            ['data/iris_sp_chars.csv', {"fields"=> {}}, {"pétal.length"=>1, "pétal&width\u0000"=> 2, "sépal.length"=>1, "sépal&width"=> 2, "spécies"=> "Iris-setosa"}, 'Cluster 7', 0.8752380218327035],
            ['data/movies.csv', {"fields"=> {"000007"=> {"optype"=> "items", "item_analysis"=> {"separator"=> "$"}}}}, {"gender"=> "Female", "age_range"=> "18-24", "genres"=> "Adventure$Action", "timestamp"=> 993906291, "occupation"=> "K-12 student", "zipcode"=> 59583, "rating"=> 3}, 'Cluster 1', 0.7294650227133437]
@@ -204,8 +204,8 @@ class Test33ComparePredictions < Test::Unit::TestCase
   # Scenario: Successfully comparing topic distributions:
   def test_scenario4
     
-     data = [[File.dirname(__FILE__)+'/data/spam.csv', {"fields" => {"000001" => {"optype": "text", "term_analysis" => {"case_sensitive" => true, "stem_words" => true, "use_stopwords" => false, "language" => "en"}}}}, {"Type" => "ham", "Message" => "Mobile call"}, [0.01878, 0.00388, 0.00388, 0.00388, 0.20313, 0.47315, 0.00574, 0.05695, 0.00388, 0.19382, 0.00388, 0.02902]],
-             [File.dirname(__FILE__)+'/data/spam.csv', {"fields" => {"000001" => {"optype": "text", "term_analysis" => {"case_sensitive" => true, "stem_words" => true, "use_stopwords" => false, "language" => "en"}}}}, {"Type" => "ham", "Message" => "Go until jurong point, crazy.. Available only in bugis n great world la e buffet... Cine there got amore wat..."}, [0.00263, 0.01083, 0.00831, 0.06004, 0.33701, 0.00263, 0.01209, 0.44553, 0.0531, 0.00326, 0.06193, 0.00263]]
+     data = [[File.dirname(__FILE__)+'/data/spam.csv', {"fields" => {"000001" => {"optype": "text", "term_analysis" => {"case_sensitive" => true, "stem_words" => true, "use_stopwords" => false, "language" => "en"}}}}, {"Type" => "ham", "Message" => "Mobile call"}, [0.51133, 0.00388, 0.00574, 0.00388, 0.00388, 0.00388, 0.00388, 0.00388, 0.00388, 0.00388, 0.00388, 0.44801]],
+             [File.dirname(__FILE__)+'/data/spam.csv', {"fields" => {"000001" => {"optype": "text", "term_analysis" => {"case_sensitive" => true, "stem_words" => true, "use_stopwords" => false, "language" => "en"}}}}, {"Type" => "ham", "Message" => "Go until jurong point, crazy.. Available only in bugis n great world la e buffet... Cine there got amore wat..."}, [0.39188, 0.00643, 0.00264, 0.00643, 0.08112, 0.00264, 0.37352, 0.0115, 0.00707, 0.00327, 0.00264, 0.11086]]
             ]
      puts
      puts "Scenario: Successfully comparing topic distributions:"
@@ -246,7 +246,6 @@ class Test33ComparePredictions < Test::Unit::TestCase
         puts "And I create a local topic distribution for <%s>" % JSON.generate(data_input)
         topic_distribution = @api.create_topic_distribution(topic_model, data_input)
         assert_equal(BigML::HTTP_CREATED, topic_distribution["code"])
-        #assert_equal(1, topic_distribution["object"]["status"]["code"])
         assert_equal(@api.ok(topic_distribution), true)
         
         puts "Then the local topic distribution is <%s>" %  JSON.generate(topic_distribution_result)